loqi-fs/requirements.txt
anyio==3.6.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.87.0
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
h11==0.14.0
httptools==0.5.0
idna==3.4
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yapf==0.32.0
